VendingMachine 관련 프로토콜 이동
- VendingMachine과 한 파일에 두어 테스트 용이하도록 변경
